grade-service/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.29
asyncpg==0.28.0
pydantic==1.10.12
python-dotenv==1.0.0
PyJWT==2.9.0
alembic==1.12.0
httpx==0.27.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.0.0
asyncpg